+++ src/disk/sys/mod.rs
@@ -1,11 +1,13 @@
-#[cfg(target_os = "linux")]
-mod linux;
+cfg_if::cfg_if! {
+	if #[cfg(target_os = "linux")] {
+		mod linux;
+		pub use linux::*;
+	}
+}
 
-#[cfg(target_os = "linux")]
-pub use linux::*;
-
-#[cfg(target_family = "unix")]
-mod unix;
-
-#[cfg(target_family = "unix")]
-pub use unix::*;
+cfg_if::cfg_if! {
+	if #[cfg(target_family = "unix")] {
+		mod unix;
+		pub use unix::*;
+	}
+}