+++ src/memory/sys/macos/mod.rs
@@ -1,0 +1,5 @@
+mod swap_memory;
+mod virtual_memory;
+
+pub use swap_memory::*;
+pub use virtual_memory::*;