+++ src/network/sys/mod.rs
@@ -1,5 +1,6 @@
-#[cfg(target_os = "linux")]
-mod linux;
-
-#[cfg(target_os = "linux")]
-pub use linux::*;
+cfg_if::cfg_if! {
+	if #[cfg(target_os = "linux")] {
+		mod linux;
+		pub use linux::*;
+	}
+}