+++ src/process/sys/macos/pids.rs
@@ -1,0 +1,11 @@
+use std::io;
+
+use crate::Pid;
+
+pub fn pids() -> io::Result<Vec<Pid>> {
+	todo!()
+}
+
+pub fn pid_exists(_pid: Pid) -> bool {
+	todo!()
+}