+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- [macos] get macos to compile
 - [cpu][all] implement cpu_count and cpu_count_physical
 - [cpu][macos] implement cpu_times, cpu_times_percent, and cpu_percent
 - [disk][unix] implement disk_usage